backend/pyproject.toml
[project]
name = "adk-agent"
version = "0.0.1"
description = "Backend for the google-adk Agent"
authors = [
    { name = "John Vandivier", email = "john@ladderly.io" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.12",
    "google-genai>=1.20.0",
    "google-adk>=1.3.0",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.2",
    "uvicorn[standard]>=0.32.1",
    "alembic>=1.16.4",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.7",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.2.0",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.11.13"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
